# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WPMgr API
  version: 0.0.0
  description: |
    Control-plane REST API for WPMgr. This is the single source of truth for
    both the Go server types and the generated TypeScript client. Real paths
    and schemas are filled in during Phase 4 (skeleton) and Phase 5 (features).
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.txt
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/sites/{siteId}/backups:
    post:
      operationId: createBackup
      summary: Start a backup of a site
      description: |
        Records a pending backup snapshot for the site and enqueues a background
        job that dispatches a signed `backup` command to the site's agent. The
        agent chunks, encrypts (client-side, age, to the site's PUBLIC recipient)
        and uploads ciphertext directly to object storage via presigned URLs,
        then submits the manifest. Incremental: a chunk whose content hash is
        already stored is not re-uploaded. Requires operator+.
      tags: [backups]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupCreate"
      responses:
        "201":
          description: Backup snapshot created (job enqueued)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupSnapshot"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listBackups
      summary: List a site's backup snapshots
      tags: [backups]
      parameters:
        - $ref: "#/components/parameters/SiteId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: A page of backup snapshots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupSnapshotList"
  /api/v1/backups/{snapshotId}:
    get:
      operationId: getBackup
      summary: Get a backup snapshot with its manifest summary
      tags: [backups]
      parameters:
        - $ref: "#/components/parameters/SnapshotId"
      responses:
        "200":
          description: The backup snapshot and its manifest entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupSnapshotDetail"
        "404":
          description: Snapshot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/backups/{snapshotId}/events:
    get:
      operationId: streamBackupSnapshotEvents
      summary: Stream live backup snapshot progress (SSE)
      description: |
        Server-Sent Events stream (text/event-stream) of progress transitions for
        a backup snapshot, for live UI updates. Each `data:` line is a JSON
        BackupEvent. The stream emits periodic heartbeat comment lines (":\n")
        and ends when the snapshot reaches a terminal state (completed/failed)
        or the client disconnects. Requires viewer+.
      tags: [backups]
      parameters:
        - $ref: "#/components/parameters/SnapshotId"
      responses:
        "200":
          description: |
            An event stream. Media type text/event-stream; each event payload is
            a BackupEvent JSON object. Documented here as the BackupEvent schema
            for clients/codegen even though the transport is SSE, not JSON.
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/BackupEvent"
        "404":
          description: Backup snapshot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/backups/{snapshotId}/restore:
    post:
      operationId: createRestore
      summary: Restore from a backup snapshot (full or partial)
      description: |
        Enqueues a restore job. The control plane resolves the (possibly partial)
        selection into ordered chunks, issues presigned GET URLs, and dispatches
        a signed `restore` command. The agent downloads ciphertext, verifies the
        BLAKE3 of each chunk, decrypts with the age identity it alone holds, and
        reassembles. Supports full, by-path, and by-db-table partial restore.
        Requires operator+.
      tags: [backups]
      parameters:
        - $ref: "#/components/parameters/SnapshotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreCreate"
      responses:
        "202":
          description: Restore job enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupSnapshot"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/sites/{siteId}/backup-schedule:
    get:
      operationId: getBackupSchedule
      summary: Get a site's backup schedule
      tags: [backups]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "200":
          description: The site's backup schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupSchedule"
        "404":
          description: No schedule configured for the site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: putBackupSchedule
      summary: Create or update a site's backup schedule
      tags: [backups]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupScheduleUpdate"
      responses:
        "200":
          description: The saved backup schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupSchedule"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
...
    SiteId:
      name: siteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SnapshotId:
      name: snapshotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    AgentSignature:
      type: apiKey
      in: header
      name: X-WPMgr-Signature
      description: |
        Ed25519 signed-request authentication for agent->control-plane calls. The
        agent sends four headers: X-WPMgr-Agent-Key (its base64 Ed25519 public
        key, identifying the enrolled site), X-WPMgr-Timestamp (Unix seconds),
        X-WPMgr-Nonce (a unique per-request token), and X-WPMgr-Signature (the
        base64 Ed25519 signature over the canonical message
        "METHOD\nPATH\nTIMESTAMP\nNONCE\nhex(sha256(body))"). The control plane
        rejects stale timestamps and replayed nonces. (apiKey type is the closest
        OpenAPI 3.1 primitive for this header-carried signature scheme.)
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-readable error description.
        details:
          type: object
          additionalProperties: true
...
    BackupEvent:
      type: object
      description: |
        One snapshot-progress transition streamed over SSE. Status mirrors the
        parent snapshot's status at publish time so a subscriber can detect a
        terminal state (completed/failed) and close the EventSource. Phase
        values come from the closed `allowedProgressPhases` set on the CP.
      required: [snapshot_id, phase, status, ts]
      properties:
        snapshot_id:
          type: string
          format: uuid
        phase:
          type: string
          enum:
            - queued
            - started
            - dumping_db
            - archiving_files
            - encrypting_uploading
            - compressing_files
            - encrypting
            - uploading
            - submitting_manifest
            - completed
            - failed
        phase_detail:
          type: object
          additionalProperties: true
          description: Pass-through of the agent's POST /progress payload (e.g. chunk counters).
        status:
          type: string
          enum: [pending, running, completed, failed]
        ts:
          type: string
          format: date-time
    BackupCreate:
      type: object
      description: Request to start a backup of a site.
      properties:
        kind:
          type: string
          enum: [files, db, full]
          default: full
          description: What to back up. Defaults to full (files + database).
    BackupSnapshot:
      type: object
      required: [id, tenant_id, site_id, kind, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        site_id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        kind:
          type: string
          enum: [files, db, full]
        status:
          type: string
          enum: [pending, running, completed, failed]
        age_recipient:
          type: string
          description: |
            The age PUBLIC recipient the chunks were encrypted to (provenance).
            NEVER a private key; the control plane cannot decrypt backups.
        total_size:
          type: integer
          format: int64
        chunk_count:
          type: integer
          format: int64
        archived:
          type: boolean
          description: Kept by the monthly-archive retention rule.
        error:
          type: string
        progress:
          type: object
          additionalProperties: true
          description: |
            M5.6 / ADR-032 phpbu runner progress. Empty `{}` until the runner posts
            its first phase. Shape:
              { "phase": "uploading", "phase_detail": {"chunks_done": 17, "chunks_total": 42, ...} }
            Phases form a closed set (see backup.allowedProgressPhases on the CP).
        progress_updated_at:
          type: string
          format: date-time
          description: |
            Server timestamp of the last progress POST from the runner. Used by the
            CP watchdog (>120s without an update on a running snapshot → marked
            failed/stalled) and by the frontend to detect a silent runner.
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BackupSnapshotList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BackupSnapshot"
...
    BackupSnapshotDetail:
      type: object
      required: [snapshot, entries]
      properties:
        snapshot:
          $ref: "#/components/schemas/BackupSnapshot"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/BackupManifestEntry"
...